Player Class; Map <-> MapsActivty
Implemented player class
Resulted in some changes  and additions elsewhere as well.
Made an instance and getInstance for MapActivity, so that the values
from there, e.g. the google map (because it can be null, an the null fix
can only be done in an Activity) so that Player can have a location
value.
This required some skeleton (with missing bones) impl. of Map.