Map, AreaGrid, Area Classes Implemented
Lots of changes. The Map has been implemented, as well as everything it
needed to run.
BlobPS has been changed to so that a BlobPS is actually made, which sets
the instance. From the command line and old BlobPS can be loaded.
Methods and other more minor changes have been made to many class.